Move readiness registration tests
